requirements.txt
# Profit Leakage & True Unit Economics Analyzer
# Python dependencies — full stack including dashboard

# Data pipeline
psycopg2-binary>=2.9.9
sqlalchemy
pandas
pyyaml==6.0.1
tabulate==0.9.0

# Dashboard
streamlit
plotly
